# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Microservice API
  version: 1.0.0
  description: Mock Payment Microservice API — authorize, capture, refund, get status.

servers:
  - url: http://localhost:8085
    description: Local dev

paths:
  /payments:
    post:
      tags: [Payments]
      summary: Authorize a payment
      operationId: authorizePayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
            examples:
              example-1:
                value:
                  amount: 120.50
                  currency: USD
                  metadata:
                    orderId: ORD-1234
      responses:
        '201':
          description: Payment authorized (or pending/declined)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment status/details
      operationId: getPayment
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /payments/{id}/capture:
    post:
      tags: [Payments]
      summary: Capture an authorized payment
      operationId: capturePayment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund a captured payment
      operationId: refundPayment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:

  schemas:

    AuthorizeRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          example: 120.50
        currency:
          type: string
          example: USD
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            orderId: ORD-1234

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        externalTransactionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string

    PaymentStatus:
      type: string
      enum:
        - PENDING
        - AUTHORIZED
        - CAPTURED
        - REFUNDED
        - DECLINED
        - FAILED

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'